feat: multi-site support for refresh token
The previous implementation of the CraftagramService would exit if the first site did not have a configured long access token. It would also return `true` if the API call to Instagram failed. The console controller returned the raw bool value as well, which mean that `true` gets converted to `1` which shells interpret as "exited with an error". Also fix some typos and unused namespace imports.